+++ src/app/app-routing.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomepageComponent } from './homepage/homepage.component';
-import { SigninComponent } from './signin/signin.component';
-import { SignupComponent } from './signup/signup.component';
+import { ProfileComponent } from './profile/profile.component';
 
 const routes: Routes = [
-  {path: 'signup', component:SignupComponent},
-  {path: 'signin', component:SigninComponent},
   {path: '', component:HomepageComponent},
+  {path: 'updateProfile', component:ProfileComponent},
 ];
 
 @NgModule({
