+++ src/app/app.component.html
@@ -1,2 +1,1 @@
-<app-homepage></app-homepage>
 <router-outlet></router-outlet>
