--- src/app/signin/signin.component.html
@@ -1,46 +1,0 @@
-<form [formGroup]="signinForm" (ngSubmit)="signin()">
-<div class="ms_register_popup">
-    <div id="myModal1" class="modal  centered-modal" role="dialog">
-        <div class="modal-dialog login_dialog">
-        <!-- Modal content-->
-        <div class="modal-content">
-        <button type="button" class="close" data-dismiss="modal">
-            <i class="fa_icon form_close"></i>
-        </button>
-        <div class="modal-body">
-            <div class="ms_register_img">
-                <img src="assets/images/register_img.png" alt="" class="img-fluid" />
-            </div>
-        <div class="ms_register_form">
-        <h2>Login / Sign in</h2>
-        <div class="form-group">
-            <input type="text" placeholder="Enter Your Username" class="form-control" formControlName="username">
-            <span class="form_icon">
-				<i class="fa_icon form-envelope" aria-hidden="true"></i>
-			</span>
-        </div>
-            <div class="form-group">
-                <input type="password" placeholder="Enter Password" class="form-control" formControlName="password">
-                <span class="form_icon">
-					<i class="fa_icon form-lock" aria-hidden="true"></i>
-				</span>
-            </div>
-        <div class="remember_checkbox">
-            <label>Keep me signed in
-                <input type="checkbox">
-                <span class="checkmark"></span>
-            </label>
-        </div>
-        <!-- <a class="ms_btn" target="_blank">login now</a> -->
-        <button type="submit">login now</button>
-        <div class="popup_forgot">
-            <a href="#">Forgot Password ?</a>
-        </div>
-            <p>Don't Have An Account? <a href="#myModal" data-toggle="modal" class="ms_modal1 hideCurrentModel">register here</a></p>
-        </div>
-        </div>
-        </div>
-    </div>
-    </div>
-</div>  
-</form>
