--- src/app/signup/signup.component.html
@@ -1,57 +1,0 @@
-<form [formGroup]="signupForm" (ngSubmit)="register()">
-<div class="ms_register_popup">
-    <div id="myModal" class="modal  centered-modal" role="dialog">
-        <div class="modal-dialog register_dialog">
-            <!-- Modal content-->
-            <div class="modal-content">
-                <button type="button" class="close" data-dismiss="modal">
-                    <i class="fa_icon form_close"></i>
-                </button>
-                <div class="modal-body">
-                    <div class="ms_register_img">
-                        <img src="images/register_img.png" alt="" class="img-fluid" />
-                    </div>
-                    <div class="ms_register_form">
-                        <h2>Register / Sign Up</h2>
-                        <div class="form-group">
-                            <input type="text" placeholder="Enter Your Name" class="form-control" formControlName="name">
-                            <span class="form_icon">
-                                <i class="fa_icon form-user" aria-hidden="true"></i>
-                            </span>
-                        </div>
-                        <div class="form-group">
-                            <input type="text" placeholder="Enter Your Username" class="form-control" formControlName="username">
-                            <span class="form_icon">
-                                <i class="fa_icon form-user" aria-hidden="true"></i>
-                            </span>
-                        </div>
-                        <div class="form-group">
-                            <input type="text" placeholder="Enter Your Email" class="form-control" formControlName="email">
-                            <span class="form_icon">
-                                <i class="fa_icon form-envelope" aria-hidden="true"></i>
-                            </span>
-                        </div>
-                        <div class="form-group">
-                            <input type="password" placeholder="Enter Password" class="form-control" formControlName="password">
-                            <span class="form_icon">
-                                <i class="fa_icon form-lock" aria-hidden="true"></i>
-                            </span>
-                        </div>
-                        <div class="form-group">
-                            <input type="password" placeholder="Confirm Password" class="form-control">
-                            <span class="form_icon">
-                                <i class=" fa_icon form-lock" aria-hidden="true"></i>
-                            </span>
-                        </div>
-                        <div>
-                            <button type="submit" class="ms_btn">register now</button>
-                        </div>
-                        <!-- <a class="ms_btn">register now</a> -->
-                        <p>Already Have An Account? <a href="#myModal1" data-toggle="modal" class="ms_modal hideCurrentModel">login here</a></p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-</form>
