+++ src/assets/js/plugins/player/volume.css
@@ -3,7 +3,7 @@
 } */
 
 body {
-  font-size: 14rem;
+  font-size: 14px;
 }
 
 * {
@@ -26,14 +26,14 @@ body {
   content: "";
   position: absolute;
   top: auto;
-  bottom: 6rem;
-  left: 6rem;
-  height: 7rem;
-  width: 7rem;
+  bottom: 6px;
+  left: 6px;
+  height: 7px;
+  width: 7px;
   background: #fff;
   border: none !important;
   border-radius: 100% !important;
-  box-shadow: 0 0 0 3rem rgba(0, 0, 255, 0.2), 0 0 2px 0 #fafafa;
+  box-shadow: 0 0 0 3px rgba(0, 0, 255, 0.2), 0 0 2px 0 #fafafa;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper {
   width: 55px;
@@ -64,10 +64,10 @@ body {
 .knob-container .knob-wrapper-outer .knob-wrapper .knob-mask:after {
   content: "";
   position: absolute;
-  top: 0rem;
-  left: 0rem;
-  right: 0rem;
-  bottom: 0rem;
+  top: 0px;
+  left: 0px;
+  right: 0px;
+  bottom: 0px;
   border-radius: 100%;
   border: 5px solid transparent;
 	border-bottom-color: grey;
@@ -75,10 +75,10 @@ body {
 .knob-container .knob-wrapper-outer .knob-wrapper .round {
     content: "";
     position: absolute;
-    top: 14rem;
-    left: 18rem;
-    right: 10rem;
-    bottom: 18rem;
+    top: 14px;
+    left: 18px;
+    right: 10px;
+    bottom: 18px;
     border-radius: 100%;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper .knob {
@@ -98,7 +98,7 @@ body {
   right: 1px;
   bottom: 1px;
   border-radius: 100%;
-  border: 3rem solid transparent;
+  border: 3px solid transparent;
   border-bottom-color: #fff;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper .knob.d2:after {
@@ -121,8 +121,8 @@ body {
 .knob-container .knob-wrapper-outer .knob-wrapper .handle:before {
   content: "";
   position: absolute;
-  bottom: 6rem;
-  left: 2rem;
+  bottom: 6px;
+  left: 2px;
   height: 14px;
   width: 14px;
   background: #fff;
@@ -132,28 +132,28 @@ body {
 .knob-container .knob-wrapper-outer .knob-wrapper .handle:after {
   content: "";
   position: absolute;
-  bottom: 0rem;
-  left: -1rem;
-  height: 20rem;
-  width: 20rem;
+  bottom: 0px;
+  left: -1px;
+  height: 20px;
+  width: 20px;
   border: none !important;
   border-radius: 100% !important;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper .pip {
   position: absolute;
-  font-size: 11rem;
+  font-size: 11px;
   color: black;
   pointer-events: none;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper .pip.min {
-  bottom: 6rem;
+  bottom: 6px;
   left: 0;
   right: 50%;
   width: 50%;
   text-align: center;
 }
 .knob-container .knob-wrapper-outer .knob-wrapper .pip.max {
-  bottom: 6rem;
+  bottom: 6px;
   left: 50%;
   width: 50%;
   right: 0;
