--- src/home1.html
@@ -1,1 +1,0 @@
-<app-root></app-root>
